Most of the tests are now passed.
The outstanding issue is to impose restrictions on the usage of types.
In particular, the new grammar allows using subtypes everywhere. For example,
it is possible to define a structure subtypes members or to create subtype
aliases (e.g. C = A.B). We should find a way to identify these cases and report
adequate errors.