Add functionality to map from linear to full positions.
The added code provides a couple of functions to create the mapping and to use
it. The implementation I came up with is moderately sophisticated, maybe an
overkill (but it was good fun). It should provide good memory usage (the size
of the map in memory should require not much more than 1 byte per line of input
source file) and good speed (it uses a hybrid binary-search-tree /
linear-search). The work on this is not fully finished, yet.