Hack to use boxer with Python rather than Box
Started to hack Boxer to use Python as the driving language. Started
also to put together a clone of the Box graphic library for Python.
For now, this clone is pushing Box's programming paradigm into Python
(objects' constructors accept an unordered list of arguments, which are
matched by type to different object methods). This seems to work to a
good extent, but it is unclear whether it is the best approach for
Python. This commit is highly experimental and messy.  Things will
becleaned up gradually in future commits.