Fix computation of bounding box
The clipping area was passed explicitly to the on_step command, which
was a bad idea as this kind of information is not well representable
by a couple of points. We now compute the bounding box internally in
the command executor, which should prevent transformation issues.
Note that this introduces an order constraint when drawing on a deep
surface: we now must draw the Window object before drawing the depth
surface, as the latter uses a clipping area which is automatically
computed from the bounding box of the former.