Allow getting point in arc
If a is an Arc, a[0] and a[1] can now be used to obtain the start and
end point of the Arc. Floating point numbers can be used to get points
in between. For example, a[0.5] gives the point in the middle of the
Arc.

All these calculations are still subject to problems when the
curvature of the arc approaches 0 (the radius becomes infinite and the
arc becomes a straight segment). For now we accept this, as Cairo itself
seems to lack the ability of tracing arcs with arbitrarily small
curvature. We may fix the problem in the future (e.g. using a Bezier
curve to approximate quasi straight arcs).