Implement drawing a pattern via the Image object
Also:

- Improve the behaviour of the Scale object (which couldn't take a
  Scale as its argument and now it does).
- Add a draw method to Window which draws the window to a Cairo
  surface and returns it.
- Implement Image so that it can take either the name of a file
  containing the image or a cairo.ImageSurface.

These changes allow to draw something to a Window, convert it to a
Cairo surface and pass the latter to an Image so that it can be used
as a fill pattern. These feature are demonstrated in a new example
which is also added in this commit with the name house.py.