Run Python scripts via multiprocessing
Rewrite the way we launch scripts in the GUI to use the multiprocessing
module rather than subprocess. This allows better communication with the
launched script. It will also remove the need of creating temporary
files to store the script and will allow the GUI to build the
environment used by the script when running.