Prototype for moving reference points from script
Prototype which allows the Python script to communicate back with the
GUI to move a reference point. This can be used to enforce constraints
on reference points (e.g. ensuring a reference point lies on a line).
An example of this was added in the examples directory, but note that
the interface for moving reference points will change in the future.