Refactor mesh posing code
The code is refactored to allow different types of constraints.
A new Constraints object allows collecting constraints with respect
to different simultaneous views of the same mesh.

At the moment only one type of constraint is implemented
(RotationConstraint), where the orientation of the bone is determined
by specifying a rotation axis and the position of the bone end in one
selected view of the mesh.