Use textures to paint 3D meshes
This commit enables the deepsurface module to use textures for
rendering 3D meshes. This commit is mostly concerned with wiring
Python and C++ together so that the textures are passed correctly and
the corresponding Python containers are bound correctly to the C++
Mesh object which uses them. At the moment each triangle is coloured
uniformly using the color in the texture at the center of the triangle.
Future commits will improve on this.