Paint triangles using associated texture
Rather than painting each triangle uniformly with the color at the
center of the triangle, we now interpolate texture coordinates for every
pixel of the triangle and extract the corresponding color from the
texture. This feature was extremely easy to implement, using the
generalisations introduced in previous commits. The texture coordinates
are just other two ScalarAttributes.