Fix mesh auto reposing for bone axis rotations
In order to rotate a bone around its own axis we currently use Tri
points (the triple of points used for defining Bezier curves) and
abuse the in-point to determine the rotation axis. A problem occurred
when transforming a normal Point to a Tri for the first time, as the
tri.gui.old_value was containing garbage. This is fixed by ensuring a
Tri point is always created by setting the in-point coordinates to be
the same as the original parent point. This allows to detect the
situation in the mesh posing code.