Do not import gui files by default
This commit prevents importing pyrtist.gui as a side effect of
doing "import pyrtist"

Rationale: pyrtist can be used as a Python module and in that case all
the stuff defined in pyrtist.gui is unused. Also pyrtist.gui pulls in
PyGTK which is not available on Python 3. This commit thus allows
using Python 3 to run pyrtist scripts, even if the GUI can only run
in Python 2 at the moment (PyGTK is not available in Python 3, AFAIK.)