gi: repair visualization of reference points
This commit makes Pyrtist usable on Python3 + gi (aka PyGObject.)

This is a minimal change with minimal fixes to the drawing of reference
points. At the moment the whole DrawableArea is repainted (queue_draw)
whenever anything changes. The old code path calculating the damage area
is not used.

The code would need a good refactoring, but this is left to separate
changes.