opengl: improve reporting of GLSL compilation messages
FragmentWindow now uses the inspect module to find the line where its
set_source() method is called and uses it to generate a #line GLSL
preprocessor directive to ensure compilation messages are easier to
address. Documentation in set_source() explains how this can be used.

Also improve the reporting of error messages by decoding the messages
to strings.