opengl: allow saving to file outside GUI (headless)
Introduce a FragmentWindow.set_vars method to set variables from a
dictionary, ignoring values that do not have a recognised type.
GUI variables can now be conveniently set from a Pyrtist script with
something like: frag_win.set_vars(locals())

Change the FragmentWindow.save() method so that it only renders and
saves the scene, without setting the variables, as this is now done
in the set_vars method.

This allows running a script and saving its output from the command
line.