Add gui.is_connected method
This method can be used to distinguish whether the script is being ran
under the Pyrtist UI or as a regular python script. E.g. a script may
want to save output to images only when running as a Python script, but
not when running under the UI (as this slows down the UI, for example.)