Fix issues with newer setuptools versions
setuptools.Feature was removed. This commit changes setup.py to
use Extension, instead.

Note that the two experimental modules pyrtist.deepsurface and
pyrtist.lib3d.perf_critical are always compiled, but are marked as
optional. This means that failures to compile these modules (e.g.
because required libraries and components are not found) should
not stop the user from building and installing the software.