opengl: add ability to use textures and uniforms
Add the ability of loading a texture and using it from the OpenGL
shader. Also add the ability of setting scalar and vector uniforms.

Also add an example showing how to use this feature. The example
also shows how textures can be used to combine 2D vector graphics
in pyrtist.lib2d with OpenGL shader-based graphics in pyrtist.opengl.